requirements.txt
feedparser==6.0.11
requests==2.32.3
jinja2==3.1.4
python-dateutil==2.9.0
openai==1.68.2
